+++ ErsatzTV.Application/Streaming/Queries/GetPlayoutItemProcessByChannelNumberHandler.cs
@@ -27,6 +27,8 @@ namespace ErsatzTV.Application.Streaming;
 
 public class GetPlayoutItemProcessByChannelNumberHandler : FFmpegProcessHandler<GetPlayoutItemProcessByChannelNumber>
 {
+    private static readonly Random FallbackRandom = new();
+
     private readonly IArtistRepository _artistRepository;
     private readonly IEmbyPathReplacementService _embyPathReplacementService;
     private readonly IExternalJsonPlayoutItemProvider _externalJsonPlayoutItemProvider;
@@ -284,9 +286,20 @@ public class GetPlayoutItemProcessByChannelNumberHandler : FFmpegProcessHandler<
 
             bool isComplete = true;
 
+            bool effectiveRealtime = request.HlsRealtime;
+
+            // only work ahead on fallback filler up to 3 minutes in duration
+            // since we always transcode a full fallback filler item
+            if (!effectiveRealtime &&
+                playoutItemWithPath.PlayoutItem.FillerKind is FillerKind.Fallback &&
+                duration > TimeSpan.FromMinutes(3))
+            {
+                effectiveRealtime = true;
+            }
+
             TimeSpan limit = TimeSpan.Zero;
 
-            if (!request.HlsRealtime)
+            if (!effectiveRealtime)
             {
                 // if we are working ahead, limit to 44s (multiple of segment size)
                 limit = TimeSpan.FromSeconds(44);
@@ -319,7 +332,7 @@ public class GetPlayoutItemProcessByChannelNumberHandler : FFmpegProcessHandler<
                     now,
                     duration,
                     $"DEBUG_NO_SYNC:\n{Mapper.GetDisplayTitle(playoutItemWithPath.PlayoutItem.MediaItem, Option<string>.None)}\nFrom: {start} To: {finish}",
-                    request.HlsRealtime,
+                    effectiveRealtime,
                     request.PtsOffset,
                     channel.FFmpegProfile.VaapiDisplay,
                     channel.FFmpegProfile.VaapiDriver,
@@ -334,7 +347,8 @@ public class GetPlayoutItemProcessByChannelNumberHandler : FFmpegProcessHandler<
                     true,
                     effectiveNow.ToUnixTimeSeconds(),
                     Option<int>.None,
-                    Optional(channel.PlayoutOffset));
+                    Optional(channel.PlayoutOffset),
+                    !effectiveRealtime);
             }
 
             MediaVersion version = playoutItemWithPath.PlayoutItem.MediaItem.GetHeadVersion();
@@ -443,7 +457,7 @@ public class GetPlayoutItemProcessByChannelNumberHandler : FFmpegProcessHandler<
                 channel.FFmpegProfile.VaapiDriver,
                 channel.FFmpegProfile.VaapiDevice,
                 Optional(channel.FFmpegProfile.QsvExtraHardwareFrames),
-                hlsRealtime: request.HlsRealtime,
+                effectiveRealtime,
                 playoutItemWithPath.PlayoutItem.MediaItem is RemoteStream { IsLive: true }
                     ? StreamInputKind.Live
                     : StreamInputKind.Vod,
@@ -465,7 +479,8 @@ public class GetPlayoutItemProcessByChannelNumberHandler : FFmpegProcessHandler<
                 isComplete,
                 effectiveNow.ToUnixTimeSeconds(),
                 playoutItemResult.MediaItemId,
-                Optional(channel.PlayoutOffset));
+                Optional(channel.PlayoutOffset),
+                !effectiveRealtime);
 
             return Right<BaseError, PlayoutItemProcessModel>(result);
         }
@@ -481,6 +496,13 @@ public class GetPlayoutItemProcessByChannelNumberHandler : FFmpegProcessHandler<
                 .MapT(pi => pi.StartOffset);
 
             Option<TimeSpan> maybeDuration = maybeNextStart.Map(s => s - now);
+
+            // limit working ahead on errors to 1 minute
+            if (!request.HlsRealtime && maybeDuration.IfNone(TimeSpan.FromMinutes(2)) > TimeSpan.FromMinutes(1))
+            {
+                maybeNextStart = now.AddMinutes(1);
+                maybeDuration = TimeSpan.FromMinutes(1);
+            }
 
             DateTimeOffset finish = maybeNextStart.Match(s => s, () => now);
 
@@ -522,7 +544,8 @@ public class GetPlayoutItemProcessByChannelNumberHandler : FFmpegProcessHandler<
                         true,
                         now.ToUnixTimeSeconds(),
                         Option<int>.None,
-                        Optional(channel.PlayoutOffset));
+                        Optional(channel.PlayoutOffset),
+                        !request.HlsRealtime);
                 case PlayoutItemDoesNotExistOnDisk:
                     Command doesNotExistProcess = await _ffmpegProcessService.ForError(
                         ffmpegPath,
@@ -545,7 +568,8 @@ public class GetPlayoutItemProcessByChannelNumberHandler : FFmpegProcessHandler<
                         true,
                         now.ToUnixTimeSeconds(),
                         Option<int>.None,
-                        Optional(channel.PlayoutOffset));
+                        Optional(channel.PlayoutOffset),
+                        !request.HlsRealtime);
                 default:
                     Command errorProcess = await _ffmpegProcessService.ForError(
                         ffmpegPath,
@@ -568,7 +592,8 @@ public class GetPlayoutItemProcessByChannelNumberHandler : FFmpegProcessHandler<
                         true,
                         now.ToUnixTimeSeconds(),
                         Option<int>.None,
-                        Optional(channel.PlayoutOffset));
+                        Optional(channel.PlayoutOffset),
+                        !request.HlsRealtime);
             }
         }
 
@@ -711,8 +736,14 @@ public class GetPlayoutItemProcessByChannelNumberHandler : FFmpegProcessHandler<
                 collectionKey,
                 cancellationToken);
 
-            // TODO: shuffle? does it really matter since we loop anyway
-            MediaItem item = items[new Random().Next(items.Count)];
+            // ignore the fallback filler preset if it has no items
+            if (items.Count == 0)
+            {
+                break;
+            }
+
+            // get a random item
+            MediaItem item = items[FallbackRandom.Next(items.Count)];
 
             Option<TimeSpan> maybeDuration = await dbContext.PlayoutItems
                 .Filter(pi => pi.Playout.ChannelId == (channel.MirrorSourceChannelId ?? channel.Id))
@@ -734,15 +765,14 @@ public class GetPlayoutItemProcessByChannelNumberHandler : FFmpegProcessHandler<
                 .Filter(ms => ms.MediaVersionId == version.Id)
                 .ToListAsync(cancellationToken);
 
-            DateTimeOffset finish = maybeDuration.Match(
-                // next playout item exists
-                // loop until it starts
-                now.Add,
-                // no next playout item exists
-                // loop for 5 minutes if less than 30s, otherwise play full item
-                () => version.Duration < TimeSpan.FromSeconds(30)
-                    ? now.AddMinutes(5)
-                    : now.Add(version.Duration));
+            // always play min(duration to next item, version.Duration)
+            TimeSpan duration = maybeDuration.IfNone(version.Duration);
+            if (version.Duration < duration)
+            {
+                duration = version.Duration;
+            }
+
+            DateTimeOffset finish = now.Add(duration);
 
             var playoutItem = new PlayoutItem
             {
@@ -752,7 +782,7 @@ public class GetPlayoutItemProcessByChannelNumberHandler : FFmpegProcessHandler<
                 Finish = finish.UtcDateTime,
                 FillerKind = FillerKind.Fallback,
                 InPoint = TimeSpan.Zero,
-                OutPoint = version.Duration,
+                OutPoint = duration,
                 DisableWatermarks = !fallbackPreset.AllowWatermarks,
                 Watermarks = [],
                 PlayoutItemWatermarks = [],