+++ ErsatzTV.Application/Streaming/Queries/GetWrappedProcessByChannelNumberHandler.cs
@@ -48,6 +48,7 @@ public class GetWrappedProcessByChannelNumberHandler : FFmpegProcessHandler<GetW
             true,
             Option<long>.None,
             Option<int>.None,
-            Option<TimeSpan>.None);
+            Option<TimeSpan>.None,
+            false);
     }
 }